chore(cleanup): split up local and remote pcm tracks
