+++ pkg/media/constants.go
@@ -1,0 +1,11 @@
+package media
+
+import (
+	"time"
+)
+
+const (
+	DefaultOpusSampleRate     = 48000
+	DefaultOpusSampleDuration = 20 * time.Millisecond
+	defaultPCMFrameDuration   = 10 * time.Millisecond
+)
